// client/src/lib/graphql/queries/AuthUrl/__generated__/AuthUrl.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { LoginType } from "./../../../globalTypes";

// ====================================================
// GraphQL query operation: AuthUrl
// ====================================================

export interface AuthUrl {
  authUrl: string;
}

export interface AuthUrlVariables {
  loginType?: LoginType | null;
}